+++ NEWS.md
@@ -20,6 +20,9 @@
 * `read_parquet()` can now read empty data frames (zero rows) written by the
   arrow package (#160).
 
+* `read_parquet()` can now read Parquet files that have zero-row row groups
+  without dictionary pages (#162).
+
 # nanoparquet 0.5.0
 
 * `append_parquet()` now gives a clear error when appending data with